+++ tests/usage/usage.sh
@@ -1,0 +1,3 @@
+#!/bin/env bash
+
+bash tests/usage/project.sh